Cleanup models, Refactor SQL statments
