+++ backend/models/sql/bug.go
@@ -1,0 +1,11 @@
+package sql
+
+import (
+	"database/sql"
+)
+
+type BugDB struct {
+	Db *sql.DB
+}
+
+var BugDb = &BugDB{}
