+++ backend/models/ticket.go
@@ -1,9 +1,5 @@
 package models
 
-import (
-	"database/sql"
-)
-
 // Customer will create this ticket, then the stuff will check it first then convert this ticket to bug
 type Ticket struct {
 	Base
@@ -37,10 +33,3 @@ func (ticket *Ticket) SolveTicket(ticketId int) error {
 
 	return nil
 }
-
-// need to do more cleanup here
-type TicketDB struct {
-	Db *sql.DB
-}
-
-var TicketDb = &BugDB{}
